+++ .github/scripts/build-posix.sh
@@ -39,19 +39,10 @@ echo "::endgroup::"
 
 echo "::group::Setting up specific dependencies"
 
-opam install -y xml-light
-
 git clone https://github.com/savonet/ocaml-xiph.git
 cd ocaml-xiph
 opam install -y .
 cd ..
-
-cd /tmp
-rm -rf ocaml-posix
-git clone https://github.com/savonet/ocaml-posix.git
-cd ocaml-posix
-opam pin -ny .
-opam install -y posix-socket.2.1.0 posix-base.2.1.0 posix-time2.2.1.0 posix-types.2.1.0
 
 cd /tmp/liquidsoap-full/liquidsoap
 
@@ -64,7 +55,8 @@ cd ..
 
 opam update
 opam remove -y jemalloc
-opam install -y tls.1.0.2 ca-certs mirage-crypto-rng cstruct saturn_lockfree.0.5.0 ppx_hash memtrace
+opam unpin -n posix-socket posix-base posix-time2 posix-types
+opam install -y tls.1.0.2 ca-certs mirage-crypto-rng cstruct saturn_lockfree.0.5.0 ppx_hash memtrace xml-light posix-socket.2.1.0 posix-base.2.1.0 posix-time2.2.1.0 posix-types.2.1.0
 
 cd /tmp/liquidsoap-full